Add Client#locations association.
It parses the locations out of the client markup and builds and assigns
them to the Client. There are no additional HTTP calls, it's all in the
client's markup.

I tried to use Virtus's strict coercion on the Client so that it would
explode if there was a problem with the location's hash, but it gave me
some trouble and I skipped it.